requirements.txt
django==3.2
flake8==4.0
python-dotenv==0.20.0
django-split-settings==1.1.0
django-debug-toolbar==3.2.4
psycopg2-binary==2.9.3
pytest==7.1.1